Refactor: Organize data layer and update user data saving logic
This commit reorganizes the data layer by moving repository implementations to the `data.datasource` package and renames `OcrDatasource` to `OcrDataRepository` for clarity. It also updates the `CameraScannerViewModel` to save user data only when the scan state is `FINISH`.

Key changes:

-   **Package Restructuring**:
    -   `MlKitOcrDatasource.kt`: Moved from `data.impl` to `data.datasource`. The interface it implements, formerly `OcrDatasource`, has been renamed to `OcrDataRepository` and moved to `domain.repository`.
    -   `OcrRepositoryImpl.kt`: Moved from `domain.repository` to `data.datasource`. It now depends on `OcrDataRepository`.
    -   `UserRepositoryImpl.kt`: Moved from `domain.repository` to `data.datasource`.
-   **DI Module Updates**:
    -   `DataModule.kt`: Updated `bindOcrDatasource` to bind `MlKitOcrDatasource` to the renamed `OcrDataRepository` interface.
    -   `RepositoryModule.kt`: Updated to bind `OcrRepositoryImpl` and `UserRepositoryImpl` from their new locations in `data.datasource`.
-   **User Data Saving Logic**:
    -   `CameraScannerViewModel.kt`: The `saveUser(newUserData)` call has been moved from the `processImage` success block to the `updateScanStateBasedOnUserData` function. Now, user data is saved only when `userDataValidator.isUserDataComplete(userData)` is true, which corresponds to the `ScanState.FINISH`.
