+++ app/src/main/java/com/example/launchcamera/data/datasource/OcrRepositoryImpl.kt
@@ -1,13 +1,14 @@
-package com.example.launchcamera.domain.repository
+package com.example.launchcamera.data.datasource
 
 import androidx.camera.core.ImageProxy
-import com.example.launchcamera.data.datasource.OcrDatasource
 import com.example.launchcamera.domain.model.UserData
 import com.example.launchcamera.domain.parser.MLKitTextParser
+import com.example.launchcamera.domain.repository.OcrDataRepository
+import com.example.launchcamera.domain.repository.OcrRepository
 import javax.inject.Inject
 
 class OcrRepositoryImpl @Inject constructor(
-    private val ocrDatasource: OcrDatasource,
+    private val ocrDatasource: OcrDataRepository,
     private val mlKitTextParser: MLKitTextParser
 ): OcrRepository {
     override suspend fun extractTextFromImage(image: ImageProxy, data: UserData): Result<UserData> =
