+++ app/src/main/java/com/example/launchcamera/domain/di/RepositoryModule.kt
@@ -1,9 +1,9 @@
 package com.example.launchcamera.domain.di
 
 import com.example.launchcamera.domain.repository.OcrRepository
-import com.example.launchcamera.domain.repository.OcrRepositoryImpl
+import com.example.launchcamera.data.datasource.OcrRepositoryImpl
 import com.example.launchcamera.domain.repository.UserRepository
-import com.example.launchcamera.domain.repository.UserRepositoryImpl
+import com.example.launchcamera.data.datasource.UserRepositoryImpl
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
