+++ app/src/main/java/com/example/launchcamera/domain/repository/OcrDataRepository.kt
@@ -1,8 +1,8 @@
-package com.example.launchcamera.data.datasource
+package com.example.launchcamera.domain.repository
 
 import androidx.camera.core.ImageProxy
 import com.google.mlkit.vision.text.Text
 
-interface OcrDatasource {
+interface OcrDataRepository {
     suspend fun recognizeText(imageProxy: ImageProxy): Result<Text>
 }
